requirements.txt
# required inside container/ for pipeline
accelerate==1.1.1
dotenv
ema-pytorch==0.7.6
geomloss==0.2.6
gower==0.1.2
optuna==4.3.0
pandas
POT==0.9.5
PyYAML==6.0.2
scikit-learn==1.5.2
sdv==1.17.1
torch==2.5.0
tomli==2.2.1
tomli_w==1.2.0
torchmetrics==1.6.0
tqdm==4.67.1
wandb==0.19.0
optuna-integration[wandb]
miceforest==6.0.3

# required to run container on aws/ in local environment 
# sagemaker==2.247.1
# sagemaker-core==1.0.42
# ipykernel==6.29.5
# ipywidgets
# dotenv

# required for analysis
# matplotlib
# seaborn